ppr-ontology/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ppr-ontology-generation"
version = "0.1.0"
description = "AAS generation pipeline for the ppr-ontology project"
requires-python = ">=3.10"
dependencies = [
  "pyyaml>=6.0",
  "pymupdf4llm>=0.0.17",
  "rdflib>=7.0.0",
  "pyshacl>=0.27.0",
  "basyx-python-sdk>=1.2.1,<2.0.0",
  "openai>=1.40.0",
  "google-genai>=0.7.0",
]

[project.scripts]
generation-run = "generation.run:main"

[tool.setuptools.packages.find]
include = ["generation*"]
